FIx crash in dev enivir
This was due to there being two classes that implement @EventBusSubscriber to load items/block/models.

This is for you NovaViper, you can now start MC in the dev environment. I could do some work on the compiler but this is easier and you will rarely need to compile DT if you do ask me.

I would advise working in  a 1.12 workspace, you will notice that there will be errors in the doggytalents/base.* ignore these unless the folder is for your specfic  version of MC. base/a is 1.9.4, base/b 1.10.2 etc.

If confused ask!