Fix conflict betwen AT and Mixin when using Performant (#255)
DT makes func_209511_p public but then Performant tries to replace the function with a private function which mixin doesn't allow.

The logic here assumes that isWet is the disjunction of isInWater & isInRain, since water takes priority over rain when determining the wet source we can deduce the valie of isInRain when isInWater is false.