Fix dogs attack creepers at lvl 0 Creeper Sweeper
The CreeperSweeperTalent logic doesn't take effect at lvl 0 since the talent object is not even present on the dog. This was a problem as the logic to stop a dog attacking creepers at lvl 0 was contained in there. This defaults dogs to not attack creepers so that the talent can override it at lvl 5.

#353