Fix dog bed no longer show the owner's name
Error in 796002fccfea1dc6b56d9d8999a4f5f2ba794807

The "name" local var is removed, which deshadow the "name" identifier,
which now pointing to the DogBedTileEntity::name which is meant to
represent the name the player tag the bed with a name tag.
The original name var's function has been replaced with the
"text" var, which is the correct one.

Closes #362