Swap condition parts to avoid warning server side
This client config value is attempted to be accessed from the server casuing: "Caused by: java.lang.IllegalStateException: Cannot get config value before config is loaded.
This error is currently only thrown in the development environment, to avoid breaking published mods.
In a future version, this will also throw in the production environment."